Initial Commitment
Thank you for downloading the revived version of Isengriff's Storytime Mod. I am both humbled and honored to have continued Isengriff's legacy and bring one of the greatest 'Rimworld' Mods ever made back to the general public.

I could not have done any of this without the support of many talented men.

This includes, but is not limited to the contributions of the following participants:

* BRATWURSTINATOR
* EVELYN
* GARBAGE DAY IS HERE TO STAY
* STEVEO.O

To all others who gave me advice, thank you.

Warmest regards,
-TURTLESHROOM

-----------------------

		Look upon my works, ye mighty, and despair.

					CREDITS
					-------

					CODERS
					------
			HEAD PROGRAMMER: SPENCER

					VOICE TALENT
					------------
				TREE FROG: JONAH