V1.4 and Future Collaboration?
https://www.github.com/Lindonrow/StorytimeOfficialMod

As exhibited in the credits of this Mod, BOTH Isengriff AND Joe Zedong said that they were going to discontinue the Mod and leave it with me.

Much to my delight, as it turns out, they did NOT decide to discontinue the Mod, and pushed an excellent update to V1.4 .

I am not sure if this was meant to be a secret, but given that I had followed every Mod entry related to the Storytime Mod for well over a year, they didn't hide it well.

Regardless, if I cannot reestablish contact, I will continue my work. Thankfully, the part that I was incapable of doing has been done by Isengriff and company.

Working on your Mod has been my passion for over a year, and I really want to join you in this project.
You know the stuff I've put and the effort I made, and I just want to belong and be a part of your work.

I tried writing to you on the Github, but I was blocked there.

I'm not mad at any of you at all, but I just don't understand.

...